refactor: add `fd_util` package
