+++ src/internal/event_loop/moon.pkg.json
@@ -3,6 +3,7 @@
     "moonbitlang/async/os_error",
     "moonbitlang/async/internal/coroutine",
     "moonbitlang/async/internal/thread_pool",
+    "moonbitlang/async/internal/fd_util",
     "moonbitlang/async/internal/time"
   ],
   "native-stub": [ "poll.c" ]
