+++ src/internal/fd_util/fd_util.mbti
@@ -1,0 +1,20 @@
+// Generated using `moon info`, DON'T EDIT IT
+package "moonbitlang/async/internal/fd_util"
+
+// Values
+fn close(Int) -> Unit raise
+
+fn pipe() -> (Int, Int) raise
+
+fn set_blocking(Int) -> Unit raise
+
+fn set_cloexec(Int) -> Unit raise
+
+fn set_nonblocking(Int) -> Unit raise
+
+// Types and methods
+
+// Type aliases
+
+// Traits
+
