+++ src/internal/fd_util/moon.pkg.json
@@ -1,0 +1,4 @@
+{
+  "import": [ "moonbitlang/async/os_error" ],
+  "native-stub": [ "stub.c" ]
+}
