+++ src/pipe/ffi.mbt
@@ -13,16 +13,7 @@
 // limitations under the License.
 
 ///|
-extern "C" fn pipe_ffi(out : FixedArray[Int]) -> Int = "moonbitlang_async_pipe"
-
-///|
 extern "C" fn get_blocking(fd : Int) -> Int = "moonbitlang_async_get_blocking"
-
-///|
-extern "C" fn set_nonblock(fd : Int) -> Int = "moonbitlang_async_set_nonblock"
-
-///|
-extern "C" fn set_blocking(fd : Int) -> Int = "moonbitlang_async_set_blocking"
 
 ///|
 extern "C" fn read_ffi(
