+++ src/pipe/moon.pkg.json
@@ -2,6 +2,7 @@
   "import": [
     "moonbitlang/async/os_error",
     "moonbitlang/async/internal/event_loop",
+    "moonbitlang/async/internal/fd_util",
     "moonbitlang/async"
   ],
   "native-stub": [ "stub.c" ]
